Tweaked badge schema and mixin, added embeds, add function to generate name from title (similar to Vendor) so moved stuff around a bit to re-use some code and simplify
