+++ src/encoded/schemas/badge.json
@@ -3,16 +3,27 @@
     "id": "/profiles/badge.json",
     "$schema": "http://json-schema.org/draft-04/schema#",
     "additionalProperties": false,
+    "identifyingProperties": ["badge_name", "uuid", "aliases"],
+    "required": ["title"],
     "mixinProperties": [
         { "$ref": "mixins.json#/schema_version" },
         { "$ref": "mixins.json#/uuid" },
         { "$ref": "mixins.json#/submitted" },
+        { "$ref": "mixins.json#/attribution" },
         { "$ref": "mixins.json#/modified" }
     ],
     "type": "object",
     "properties": {
         "schema_version": {
             "default": "1"
+        },
+        "lab": {
+            "default": "828cd4fe-ebb0-4b36-a94a-d2e3a36cc989",
+            "permission": "import_items"
+        },
+        "award": {
+            "default": "b0b9c607-f8b4-4f02-93f4-9895b461334b",
+            "permission": "import_items"
         },
         "status": {
             "title": "Status",
@@ -24,31 +35,48 @@
                 "deleted"
             ]
         },
+        "title": {
+            "title": "Badge Name",
+            "description": "The display name for the badge.",
+            "type": "string",
+            "lookup": 20
+        },
         "badge_name": {
-            "title": "Name",
-            "description": "The display name for the badge",
-            "type": "string"
+            "description": "DON'T SUBMIT, auto-generated, use for referencing vendors in other sheets.",
+            "type": "string",
+            "pattern": "^[a-z0-9\\-]+$",
+            "uniqueKey": true,
+            "exclude_from": ["submit4dn", "FFedit-create"]
         },
         "badge_classification": {
             "title": "Badge Classification",
             "description": "The classification of a badge",
             "type": "string",
-            "comment": "Not sure this should be an enum or if this is even necessary",
             "enum": [
                 "WARNING",
                 "KUDOS",
                 "INTERNAL"
             ]
         },
+        "badge_icon": {
+            "title": "Badge Icon",
+            "type": "string",
+            "description": "The unicode character(s) for an icon to be displayed to indicate a badge for an item"
+        },
         "description": {
             "title": "Description",
             "type": "string",
             "formInput": "textarea"
         },
-        "badge_icon": {
-            "title": "Badge Icon",
-            "type": "string",
-            "description": "The unicode character(s) for an icon to be displayed to indicate a badge for an item"
+        "foursight_checks": {
+            "title": "Foursight Checks",
+            "description": "The foursight checks associated with this badge",
+            "type": "array",
+            "items": {
+                "title": "Foursight Check",
+                "description": "Name of a foursight check.",
+                "type":  "string"
+            }
         }
     }
 }
