+++ src/encoded/tests/data/inserts/badge.json
@@ -1,6 +1,6 @@
 [
     {
-        "badge_name": "Bathtub",
+        "title": "Bathtub",
         "badge_classification": "KUDOS",
         "description": "Given for exemplary data sets",
         "badge_icon": "&#xf2cd;",
