+++ src/encoded/types/__init__.py
@@ -36,7 +36,7 @@ class AnalysisStep(Item):
 
 @collection(
     name='badges',
-    unique_key='badge:badgename',
+    unique_key='badge:name',
     properties={
         'title': 'Badges',
         'description': 'Listing of badges for 4DN items',
@@ -46,6 +46,13 @@ class Badge(Item):
 
     item_type = 'badge'
     schema = load_schema('encoded:schemas/badge.json')
+    name_key = 'name'
+
+    def _update(self, properties, sheets=None):
+        # set name based on what is entered into title
+        properties['badge_name'] = self.set_namekey_from_title(properties)
+
+        super(Badge, self)._update(properties, sheets)
 
 
 @collection(
@@ -180,3 +187,25 @@ class Sysinfo(Item):
     schema = load_schema('encoded:schemas/sysinfo.json')
     name_key = 'name'
     embedded_list = []
+
+
+@collection(
+    name='vendors',
+    unique_key='vendor:name',
+    properties={
+        'title': 'Vendors',
+        'description': 'Listing of sources and vendors for 4DN material',
+    })
+class Vendor(Item):
+    """The Vendor class that contains the company/lab sources for reagents/cells... used."""
+
+    item_type = 'vendor'
+    schema = load_schema('encoded:schemas/vendor.json')
+    name_key = 'name'
+    embedded_list = ['award.project']
+
+    def _update(self, properties, sheets=None):
+        # set name based on what is entered into title
+        properties['name'] = self.set_namekey_from_title(properties)
+
+        super(Vendor, self)._update(properties, sheets)
