+++ src/encoded/types/antibody.py
@@ -37,9 +37,6 @@ class Antibody(Item):
         if properties.get('antibody_product_no'):
             abid = abid + ' ' + properties['antibody_product_no']
         abid = re.sub(regex, "-", abid)
-        # abid = re.sub('-', ' ', abid)
-        # abid = ''.join(ch for ch in abid if ch not in exclude)
-        # abid = re.sub(r"\s+", '-', abid)
         properties['antibody_id'] = abid
         super(Antibody, self)._update(properties, sheets)
