+++ src/encoded/types/base.py
@@ -33,6 +33,8 @@ from snovault.schema_utils import SERVER_DEFAULTS
 from jsonschema_serialize_fork import NO_DEFAULT
 
 from datetime import date
+import string
+import re
 
 
 @lru_cache()
@@ -349,6 +351,16 @@ class Item(snovault.Item):
             if 'admin' in user.properties['groups']:
                 return True
         return False
+
+    def set_namekey_from_title(self, properties):
+        name = None
+        if properties.get('title'):
+            exclude = set(string.punctuation)
+            name = properties['title']
+            name = ''.join(ch for ch in name if ch not in exclude)
+            name = re.sub(r"\s+", '-', name)
+            name = name.lower()
+        return name
 
     def _update(self, properties, sheets=None):
         props = {}
